refactor:#156 디자인 반영 및 컴포넌트 재분리
